fix: show profile edits after editing
